+++ backend/src/main/kotlin/com/woowahan/campus/announcement/domain/MessageSendEvent.kt
@@ -3,7 +3,7 @@ package com.woowahan.campus.announcement.domain
 data class MessageSendEvent(
     val announcement: Announcement,
 ) {
-    val channelId = announcement.slackChannelId.toLong()
+    val channelId = announcement.slackChannelId
     val author = announcement.author.author
     val content = announcement.content.content
 }
